removed landscape support in the deo project and changed log information to support 64 bit

+++ RBMenuBarDemo/ULViewController.m
@@ -51,7 +51,7 @@
     
     [self.menu dismissMenuAnimated:YES WithCompletionHandler:^(BOOL completed){
         
-        NSLog(@"%d index pressed", index);
+        NSLog(@"%lu index pressed", (unsigned long)index);
     }];
